Change namespace handling
