+++ kube/option.go
@@ -1,11 +1,9 @@
 package kube
 
 import (
-	"os"
 	"strings"
 
 	prompt "github.com/c-bata/go-prompt"
-	"github.com/c-bata/go-prompt/completer"
 )
 
 func optionCompleter(args []string, long bool) []prompt.Suggest {
@@ -128,50 +126,4 @@ func optionCompleter(args []string, long bool) []prompt.Suggest {
 var optionHelp = []prompt.Suggest{
 	{Text: "-h"},
 	{Text: "--help"},
-}
-
-/* Option arguments */
-
-var yamlFileCompleter = completer.FilePathCompleter{
-	IgnoreCase: true,
-	Filter: func(fi os.FileInfo) bool {
-		if fi.IsDir() {
-			return true
-		}
-		if strings.HasSuffix(fi.Name(), ".yaml") || strings.HasSuffix(fi.Name(), ".yml") {
-			return true
-		}
-		return false
-	},
-}
-
-func getPreviousOption(d prompt.Document) (cmd, option string, found bool) {
-	args := strings.Split(d.TextBeforeCursor(), " ")
-	l := len(args)
-	if l >= 2 {
-		option = args[l-2]
-	}
-	if strings.HasPrefix(option, "-") {
-		return args[0], option, true
-	}
-	return "", "", false
-}
-
-func completeOptionArguments(d prompt.Document) ([]prompt.Suggest, bool) {
-	cmd, option, found := getPreviousOption(d)
-	if !found {
-		return []prompt.Suggest{}, false
-	}
-	switch cmd {
-	case "get", "describe", "create", "delete", "replace", "patch",
-		"edit", "apply", "expose", "rolling-update", "rollout",
-		"label", "annotate", "scale", "convert", "autoscale", "top":
-		switch option {
-		case "-f", "--filename":
-			return yamlFileCompleter.Complete(d), true
-		case "-n", "--namespace":
-			return getNameSpaceSuggestions(), true
-		}
-	}
-	return []prompt.Suggest{}, false
 }
