+++ main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"fmt"
+	"os"
 
 	prompt "github.com/c-bata/go-prompt"
 	"github.com/c-bata/go-prompt/completer"
@@ -15,13 +16,19 @@ var (
 )
 
 func main() {
+	c, err := kube.NewCompleter()
+	if err != nil {
+		fmt.Println("error", err)
+		os.Exit(1)
+	}
+
 	defer debug.Teardown()
 	fmt.Printf("kube-prompt %s (rev-%s)\n", version, revision)
 	fmt.Println("Please use `exit` or `Ctrl-D` to exit this program.")
 	defer fmt.Println("Bye!")
 	p := prompt.New(
 		kube.Executor,
-		kube.Completer,
+		c.Complete,
 		prompt.OptionTitle("kube-prompt: interactive kubernetes client"),
 		prompt.OptionPrefix(">>> "),
 		prompt.OptionInputTextColor(prompt.Yellow),
